Streamline manifest output format (#22)
Update the manifest data model to include `HelmRelease` `values`. The
manifest is compacted when persisted, and now the same object can be
used for a raw release, or just persisted manifest.

Increase a breaking change to update all objects to be pyndantic models
to simplify how exclusion works so that it can work on nested objects.