Fluxtomize resources without a kustomization present (#80)
Flux allows automatically creating a `kustomization.yaml` on the fly
when there isn't one. This replicates that behavior by:
- including yaml files directly in the output
- Running `kustomize build` on subdirectories that do have a
kustomziation file

Introduce a new task abstraction to allow piping data between tasks that
are not just command line outputs. The fluxtomization does the
equivalent work of a kustomize build but in a mix of python and cli, and
the new abstraction allows us to put it at the start of the chain.

The testdata file in cluster2/apps that was removed was previously added
as a stopgap and now it is no longer needed, matching the original
layout of the cluster.

Issue #78