Migrate tests to syrup (#393)
Migrate tests to syrup, avoiding use of pytest-golden which hasn't been
updated for awhile and uses yaml files which are harder to use. The test
commands have been moved into the test fixture parameters which isn't so
bad to maintain, and makes it easy to see all commands run in one place.

The test files were grouped in some small groupings that make it
reasonable to manage the separate ambr files, since everything in one
file might be too hard to deal with.