Add support for HelmRelease value references (#472)
Add support for `HelmRelease` values references
https://fluxcd.io/flux/components/helm/helmreleases/#values-references
via the `valuesFrom` to populate the helm template. The spec is:

---
`.spec.valuesFrom` is an optional list to refer to `ConfigMap` and
`Secret` resources from which to take values. The values are merged in
the order given, with the later values overwriting earlier, and then
[inline
values](https://fluxcd.io/flux/components/helm/helmreleases/#inline-values)
overwriting those.

An item on the list offers the following subkeys:
- `kind`: Kind of the values referent, supported values are ConfigMap
and Secret.
- `name`: The .metadata.name of the values referent, in the same
namespace as the HelmRelease.
- `valuesKey` (Optional): The .data key where the values.yaml or a
specific value can be found. Defaults to values.yaml when omitted.
- `targetPath` (Optional): The YAML dot notation path at which the value
should be merged. When set, the valuesKey is expected to be a single
flat value. Defaults to empty when omitted, which results in the values
getting merged at the root.
- `optional` (Optional): Whether this values reference is optional. When
true, a not found error for the values reference is ignored, but any
valuesKey, targetPath or transient error will still result in a
reconciliation failure. Defaults to false when omitted.
---

- [x] Add test repository
- [x] Replace with an existing object/configmap
- [x] Add support for valuesKey
- [x] Add support for targetPath
- [ ] Figure out what to do about objects that don't exist (dummy
values)

#467