Remove unnecessary slugify dependency (#849)
Perhaps it's used in github actions eg by cruft, but the actual package
doesn't need this so let's run lean.


#### Research
I looked at some git history and this was an early addition which I
think is not used by anything. I didn't check if anything else needs it.
I guess pytest shouldn't really be there either but that's a different
story. I have not really ran the github actions in my fork but the tests
in this repo should surface any issues I guess!
The motivation behind removing slugify is because it is installing
unidecode which is GPL, so it restricts commercial usage. Feel free to
close if I am wrong about this!