feat(helm): add support for targetNamespace (#866)
This is a PoC PR to fix #865.

I've thought to add some tests, but, from what I see, contents of
`test_data` is reused across many tests, so any changes to one
`HelmRelease` manifest would have to be propagated across many unrelated
tests. To make sure I don't necessarily cause a big diff, I'd like to
ask what the best way to move forward would be (Which files to pick for
the tests? Should an additional `HelmRelease` be introduced or better to
modify an existing one?).

---------

Signed-off-by: Igor Beliakov <demtis.register@gmail.com>